Removes warning about double to float conversion

+++ CPU8085/basic/prototypes/include/common.cpp
@@ -262,6 +262,6 @@ void stringToFloat(const BYTE *currIn, float &number, int &length)
 		number = -number;
 	}
 
-	number *= pow(10, Exp+ExpInt);
+	number *= (float)pow(10, Exp+ExpInt);
 
 }
